[HUB] refactor ConnectionMap into its own file
